refactor: aggregate as plugin📦
