+++ packages/core/src/editor/index.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useMemo } from 'react';
-import { createEditor, Element, Editor as SlateEditor, Range } from 'slate';
+import { createEditor } from 'slate';
 import { Slate, withReact, ReactEditor } from 'slate-react';
 import { Config, CustomElement } from '../config';
 import { Editable } from '../editable';
@@ -17,6 +17,11 @@ import { Theme } from '../theme';
 import { Toolbar } from '../toolbar';
 import { debounce } from '../utils';
 import { styles } from './index.css';
+import {
+  withInsertBreak,
+  withOriginalIsVoid,
+  withOriginalNormalizeNode,
+} from './plugins';
 
 export const EditorClassName = styles.root;
 
@@ -44,72 +49,12 @@ export const Editor: React.FC<EditorProps> = ({
   onChange = () => {},
 }) => {
   const editor = useMemo(() => {
-    const editor = withReact(createEditor());
-
-    // Set domain specific constraints.
-    // @see: https://docs.slatejs.org/concepts/11-normalizing
-    const originalNormalizeNode = editor.normalizeNode;
-    editor.normalizeNode = (entry) => {
-      const [node] = entry;
-      if (Element.isElement(node)) {
-        const { elements } = config;
-        const element = elements.find((element) => {
-          return element.type === node.type;
-        });
-        if (element && element.normalizeNode) {
-          const isToReturn = element.normalizeNode(editor, entry);
-          if (isToReturn) {
-            // @see: https://docs.slatejs.org/concepts/11-normalizing#multi-pass-normalizing
-            return;
-          }
-        }
-      }
-      // Fall back to the original `normalizeNode` to enforce other constraints.
-      originalNormalizeNode(entry);
-    };
-
-    // Set non-void or void setting.
-    const originalIsVoid = editor.isVoid;
-    editor.isVoid = (elementNode) => {
-      const { elements } = config;
-      const element = elements.find((element) => {
-        return element.type === elementNode.type;
-      });
-      if (element && element.isVoid) {
-        return true;
-      }
-      return originalIsVoid(elementNode);
-    };
-
-    // @see: https://docs.slatejs.org/api/nodes/editor#insertbreak-greater-than-void
-    const originalInsertBreak = editor.insertBreak;
-    editor.insertBreak = () => {
-      const { selection } = editor;
-      if (!selection || !Range.isCollapsed(selection)) {
-        originalInsertBreak();
-        return;
-      }
-      const [match] = SlateEditor.nodes(editor, {
-        match: (n) => !SlateEditor.isEditor(n) && Element.isElement(n),
-        mode: 'lowest',
-      });
-      if (match) {
-        const [node] = match;
-        if (Element.isElement(node)) {
-          const { elements } = config;
-          const element = elements.find((element) => {
-            return element.type === node.type;
-          });
-          if (element && element.insertBreak) {
-            const isToReturn = element.insertBreak(editor, match);
-            if (isToReturn) {
-              return;
-            }
-          }
-        }
-      }
-      originalInsertBreak();
-    };
+    const baseEditor = withReact(createEditor());
+    const editor = Array.from([
+      withInsertBreak,
+      withOriginalIsVoid,
+      withOriginalNormalizeNode,
+    ]).reduce((acc, plugin) => plugin(acc, config), baseEditor);
 
     return editor;
   }, [config]);
